levels: map09 consistency issues.
Marked the eastern warp closet as hidden from automap.

Was reminded that this warehouse area is part of a map called "Datacenter" with computers everywhere. An early attempt to intentionally misalign some textures to make everything look a little more janky for that "office space filled with ancient unsorted boxes of legal and accounting whatever that no one ever reads but we can't ever throw away" look that the original has was unsuccessful because of the scale of the main warehouse area, which in turn (combined with the size and prominence of this thing and its relative proximity to the major active areas of the facility) suggests there was stuff regularly going through here and there should be an ongoing digitization project going on with things coming in.

Which implied there should be at least one scanning station on the warehouse floor itself, cleared for relatively low-security items, just to even begin to attempt to mitigate the worst of the backlog.

It's placed right by the entrance to the security/administration area, which recontextualizes the boxes up there as some workers' passive-aggressive attempt at mitigating the constant surveillance of their job.