levels: minor map09 fixes.
Southeastern trilobite would only be alerted if the player went all the way to try to uncover the southeastern secret. It is now right in the corner so it always gets LOS when the player hits the second teleport door switch.

Fixed some textures around the death pit escape.

Above said escape, the switch room crack was causing some slime trails and had to be moved around a bit.