levels: redo Map09 warehouse area. (#972)
* levels: redo Map09 warehouse area.

I've complained about the texture alignment of the crates before. I tried to fix it up, but then I got to the smaller dark areas and realized the longest parts of the map were a great big case of terminal One-Sector Twisty Dark Maze Syndrome and gave it up in favour of trying to recapture the interesting ideas behind this area with a brand new implementation.

- A big warehouse area that has you take a specific path at first, but then opens up after you get the key.
- The key is on the east end, behind a big door, in a big square thing. (Not quite the key itself now, but necessary to reach the key.)
- The way back involves going through some narrower, darker areas to the north and south.
- There are four secrets around the darker areas. One berserk, one plasma, one rockets, and one armour+backpack from a guy who died tragically alone in a dark storage area.
- Fairly constant shooting.

The whole place feels a lot less like Doom Containment Area which is a nice accidental plus.

One small change outside of the warehouse: the monster closets by the hallway no longer open while you're in the warehouse area, but only while you're already walking down towards the exit. (I've also moved the fake door in the final room to line up with some textures better, and replaced the textures on the stray crates in that T-corridor, but those have no effect on gameplay.)

* levels: tweak new map09 warehouse.

Fixed a floaty cell on the forklift.

Broke up some of the monotony in the northernmost aisle. The dirty crate blocking the door is now mysteriously moved back up to the rack rather than disappeared entirely.

Also tweaked the textures in that first grey T-corridor: gave the sides something with a seam that at least kindasorta lines up with the near-door seams, and got rid of the dark lines on the diagonal crate.

* levels: fix map09 co-op softlock.

If the bars are raised and everybody in the warehouse dies before getting the yellow key, everybody will be stuck outside. A teleporter now opens up just outside the warehouse that warps into the main intake room.

* levels: more map09 tweaks.

Since that path has a hurtfloor *and* potentially leads to a megaarmour secret it makes no sense to put armour bonuses there, so they are health now. Also gave a stimpack because hurtfloor.

Restored plain grey to the compy T-corridor, but used the seamless grey for the surrounding columns so there aren't any weird broken line bits. The computers themselves have been moved around and sculpted a bit so they align and no panels (slots?) are twisted into inexplicably inaccessible right angles.

* levels: improve Map09 aesthetics.

Add a trim to the walkable area for a better transition to the tekwall sides.

Adjust some of the tekwall greeble.

Reduce some of the colour clashes in the room.

Fix some texture alignment issues in that cross-corridor linking the red key areas. (There were some stuttering repeats and large objects being cut off in the tekwalls)

* levels: fix tutti frutti in map09 red key room.

Inappropriate TEKWALL3 by the door.

Also adjusted the greebles around the door for some pretence at symmetry.

* levels: more map09 texture adjustments.

The first door didn't quite show the entire orange-to-green gradation.

The minigun pillar had a floor texture that wasn't from that room nor could be explained as part of the red rock. It is now red rock.

That grey T-corridor has had further adjustments. That grey nothing by the armour bonuses is now more terminal/rack; the curved panels have been straightened; the technogreeble across them has been given some more variety.

* levels: map09 consistency issues.

Marked the eastern warp closet as hidden from automap.

Was reminded that this warehouse area is part of a map called "Datacenter" with computers everywhere. An early attempt to intentionally misalign some textures to make everything look a little more janky for that "office space filled with ancient unsorted boxes of legal and accounting whatever that no one ever reads but we can't ever throw away" look that the original has was unsuccessful because of the scale of the main warehouse area, which in turn (combined with the size and prominence of this thing and its relative proximity to the major active areas of the facility) suggests there was stuff regularly going through here and there should be an ongoing digitization project going on with things coming in.

Which implied there should be at least one scanning station on the warehouse floor itself, cleared for relatively low-security items, just to even begin to attempt to mitigate the worst of the backlog.

It's placed right by the entrance to the security/administration area, which recontextualizes the boxes up there as some workers' passive-aggressive attempt at mitigating the constant surveillance of their job.

* levels: map09 fixes.

Last update actually *broke* the pinkie warp closet since SLADE misguessed again which sector should be which when adding details.

Replaced octaminator with more serpentipedes on top of the crates as the faster movement was getting the octaminator stuck.

Blursphere moved forwards a bit to keep the sprite from clipping into the wall.

* levels: move map09 yellow key.

Put it a little further back to prevent a potential bump-grab exploit.

Also changed the strobe lighting around the red key.

* levels: minor map09 fixes.

Southeastern trilobite would only be alerted if the player went all the way to try to uncover the southeastern secret. It is now right in the corner so it always gets LOS when the player hits the second teleport door switch.

Fixed some textures around the death pit escape.

Above said escape, the switch room crack was causing some slime trails and had to be moved around a bit.