levels: make first major door in e1m7 obvious.
That big silver pillar was just enough to make  that door unrecognizable to me from a distance. It's poorly lit and surrounded by much more eye-catching details - half the time I manage to find the minigun secret, then run around the room half a dozen times wondering what to do next.

The corpse just behind the very first door has also been moved to draw the player's eye towards that door as well, which also helps move it away from the player corpse in the starting room.

Also fixed the HOM in the first lift into the underground maze, as well as changing the eye-catching thing into something a little bit less of an eyesore compared to its surroundings.