levels: improve e1m7 signposting. (#981)
* levels: improve e1m7 signposting.

Nearly all remote doors with switches more than a room away are now either the big brown door, the big AGM door, or the small tan door with the tan stripe going down the middle.

The first lift into the maze now has something eye-catching on the side.

Fenced-in area around blue key no longer has any invisible blocklines. You can jump down into the nukage if you want - the sequence break won't trap you, even if you fail to pick up the blue key here, and there's no reason in principle it shouldn't be considered a legit way to get through this 3-key, only-yellow-obligatory-for-exit map.

The bars around the armour near the blue key are now the correct silver bars.

The switch to lower the yellow key is now much, much more visible.

The red carpet switch that opens up the exit room is now turned around so you are facing the general direction of the door that it opens.

The windows around (1240,2320) now use a glass texture instead of being invisible and impassable.

That vent shaft bit is now dark, and its doors are no longer the "this is not a real door" texture.

The lift at the northern centre now has its down button right by the platform instead of some ways down the hall from it.

* levels: make first major door in e1m7 obvious.

That big silver pillar was just enough to make  that door unrecognizable to me from a distance. It's poorly lit and surrounded by much more eye-catching details - half the time I manage to find the minigun secret, then run around the room half a dozen times wondering what to do next.

The corpse just behind the very first door has also been moved to draw the player's eye towards that door as well, which also helps move it away from the player corpse in the starting room.

Also fixed the HOM in the first lift into the underground maze, as well as changing the eye-catching thing into something a little bit less of an eyesore compared to its surroundings.

* levels: fully jump-proof e1m7 secret.

The jumpable parts no longer count as secret. Removed a bad secret tag on the new bars.

The platform the minigun is on had to be expanded to ensure the player would be on it before grabbing the gun.