levels: change map01 secret sector tag.
The secret tag was on the door sector itself, which made it easy to accidentally jump over it via the raised platform in front of the door so it wouldn't be triggered properly until you were on your way out. The secret tag is now on the little corridor instead.