genmidi: Disable OPL2 waveform checks.
Pretty much every Sound Blaster-compatible card used an OPL3 chip
(everything since the Sound Blaster 2 in October 1991) so it seems
reasonable to add this as a requirement. Because of the way that the new
waveforms are arranged in OPL3, it's a pretty graceful degradation if
they're used on an OPL2 anyway (see datasheet, page 14).