genmidi: Import DMXOPL instruments.
This is ConSiGno's OPL instrument patch table, found here:
<https://github.com/sneakernets/DMXOPL>

This is the result of many years' development and probably the
highest quality Doom OPL instrument set that's ever been developed. I've
obtained permission from ConSiGno to submit it to Freedoom, with the
following small omissions from the upstream version:

* Instrument 107 (Shamisen)
* Percussion 36 (Bass Drum 1)
* Percussion 69 (Cabasa)
* Percussion 76 (Hi Wood Block)
* Percussion 77 (Low Wood Block)

In doing due diligence I found that the patch data for these instruments
matched the patch data from Doom's GENMIDI table. Just to be
meticulously safe we'll keep the old versions of these (nothing, in the
case of the Cabasa). Other than these, every other instrument has been
replaced with a new patch.