levels: various minor fixes. (#990)
* levels: various fixes.

E1M2
Fixed some missing and misaligned textures around the secret lowering computer block.

E2M5
The red key bars could both be squeezed through and jumped around on a jump-enabling sourceport. The red key now raises the stairs instead, and the final door before the exit also requires a red key.

E2M7
Added one more bar in front of the exit to prevent squeeze glides.

E3M2
Expanded yellow key pillar so it would match the flat.

E3M6
Untagged the skylight inside the secret as being itself a secret.

E4M1
That line of health and armour bonuses just inside the big building would have a few floaters. They've all been moved forward and should be fully clear of that ledge.

Map05
The pit with the central column of water and the spectres was being blocked off by said spectres before the player could jump into it. There is now a monster blocking line right in front of the ledge, barely far enough to guarantee space for you to drop but close enough to let the pinkies bite.
You can now freely move through the gaps between any of the objects in the blue key room.
The exit from the blue key building now has a lift instead of stairs, and the wall prevents the nearest pinkies from seeing you until you've actually dropped inside. Hopefully this at least mitigates the problem of being stuck on the stairs with a pinkie right below your ability to target.
Miscellaneous aesthetic tweaks.

Map07
The lift overlooking the big arena was still suffering from monsters blocking from below. It has been moved to the side with an intervening ceiling.
The central grey aquatex nukage trim hub was a cross shape and the exits all had very strong lines turning right, making it look like a swastika. The southern section has been tweaked to throw off the symmetry and avoid that impression.
Made it possible to move all the way around the stack of crates in the eastern area with the switch.
Miscellaneous texture improvements.

Map08
The exit bars are now impossible to squeeze glide through. They are also styled in the same manner as the moving pistons by the big switches, giving the suggestion that they would be moved in a similar manner.

Map09
Fixed some floating pickups at the collapsed roof, the red key room, and the stairs to the warehouse admin area.
Shotgun guys in the admin area switch room are now situated between the deep cracks that were preventing their movement.

Map12
Turned the two wall-facing enemies around near (2016,96).
Added a second trigger line for the minigunner ambush store room so it happens even when you approach it from the other side.
Added a second switch for the bars normally opened before reaching the storage room ambush.
Replaced the BFG secret with a megasphere since you already get a BFG in Map11.

Map18
There were some lamps on the lower floor of the southwest room that were invisible from the higher floor but would block your movement. All major blockers have been deleted or moved to the edge of the room.
The tree near the red armour to the right of the start has been moved for similar reasons.

Map20
Moved some pickups (and the dead body) away from the bottom sides of the steps to reduce the chances of someone bypassing them while moving downwards.
Made the armour/health bonus placement more unambiguous around the cross and star.
Simplified the bridge around the red key so that the rising portion is only one sector.

Map22
Because this map has been disproportionately harder to pistol start than anything near it...
Moved all 4 player starts into the starting elevator to avoid a hot start.
Flagged some ammo pickups as appearing on all skill levels, as well as the southern super shotgun and the starting area armour. The easy-only SSG in the starting area is now a chaingun and appears on all skill levels.
The starting shotgun is now only available on easy, with two shotgunners (one of them guarding the armour) in its place on medium and hard.
Differentiated the walls of the starting elevator so you know which way to face, and made the switch resemble the normal (but broken) SW1TEK.
The switch system itself is simplified and given a simpler implied story: the big switch is broken, so you locate and use the backup system. The switch shooting is gone.
Made the fringe around the lava elevator go all the way around and populated it with health bonuses to make it clear you're actually expected to go down there, because frankly the original is so thoroughly obscured it looks like you're breaking sequence by exploiting an oversight in the map.
Extended the staircase down from the brown platform to mitigate the "pinkie right in front of you below your aim" effect.
Added a medikit near that brown platform.
The armour in the crate hall is now a medikit.

Map24
The hanging bodies in the red key room are now the non-blocking versions.
The trilobite stuck on the teleport pad in the square maze is now an octaminator so it doesn't get stuck.
The lines around the door leading to the serpentipedes with their backs to you no longer block sounds.
Removed the sound block flags on the octaminator ambush doors, so the pinkies can join in if you start shooting at them instead of powering through into the new room.

Map26
Health bonuses under crusher were floating.

Map27
Lizardbaby platform now has a full 72 unit clearance.
Spectres in the red rock area to the northeast are now in the tunnels and only come out when you land in the red water.

* levels: fix one screwy texture.

Monitor in the room west of the westmost shiny lowering shelf thing.

* levels: more minor fixes.

E2M1
Life surge secret was marked on the thin doorway, making it possible to fail to register even after you've taken the surge. The secret is now the room itself, at the cost of that random light effect.

E2M6
Life surge secret was marked on the thin central bit of the sigil, leaving *lots* of room to step around it while grabbing the powerup. (The fact that the red makes it kinda look like a hurtfloor really doesn't help.) The secret is now the larger room itself.

Map12
New eastern storage room trigger wasn't covering the entire hallway. It should also be diagonal to minimize the chance of it being skipped entirely.

Map28
The serps on the ledge by the bloodfalls and fleshy sigil were facing the wall and marked as ambush. They now face the actual play area.

Map29
Moved all the secret pit pickups a little bit further inside the square so they look like they're resting on the surface of the solid floor.

Map30
The multiplayer-only spawn pickups are now flagged as multi-only.

* levels: tiny map04 aesthetic tweak.

The fake contrast exaggerates the shadow on some "AGM" silver columns while eliminating the shadow on others. These are often right next to each other on the screen, producing absurd results. This moves 1 vertex from each affected line 2 pixels so they are no longer orthogonal.