levels: minor tweaks/fixes. (#1038)
* levels: minor tweaks/fixes.

E2M1
The little pass-over route in the north with the medikit now blocks monsters so you don't get a surprise invisible obstruction.
The armour bonuses on the steps in the main path between the outdoor areas are now closer to the ledge. You can still miss some if you charge down the steps at full speed but it only takes a slight slowdown to get them all in one go.

Map11
The health bonus in the starting room pillar is replaced in hard mode by a shellbox.
Untagged the thin panel armour bonus secret.
Replaced the evil eyes in the SKAG room with skull rocks.
Made all 3 doors in the starting room green.
Fixed texture alignment on castle corpse hanger lift.

Map09
Flagged some secret area entrance lines as secret.

* levels: fix E1M1 texture stuff.

The fake barred doors in the dark halls were missized and tutti-fruttied.

The use-to-lower function of the crate has been removed since you can easily jump onto it from the lift and it never really made sense.

The two sets of computer banks have been reversed so that the 64-tall one is on top, as trying to accommodate it was resulting in weird texture misalignments elsewhere.

* levels: raise light fixture in E2M3 red key room.

This one's a bit petty maybe, but it bothers me that you can run onto those terminals just fine, but the light fixture is just low enough that once you're on the terminals you can't jump off of them without crouching.

The diegetically "normal" solution would be to raise the terminals to 25 units off the floor, but making the player go around to reach the red key - when in a typical game you will have already killed any monster that has line of fire on you anywhere in that room - adds nothing to gameplay whatsoever.