dehacked: remove armour blinking.
The single most annoying creative constraint in creating an armour sprite is to explain why it's a blinking light. It never made sense in id and the explanation looks ridiculous in the new sprites in the recent devbuilds. (I say this as the person who added them - it's still in my view preferable to the alternative pure arbitrariness.)

If we get rid of the fullbright, we can design armour like any other mundane tech/ammo pickup.