levels: Map19 fixes. (#1071)
Originally meant to address #1070.

Redid green armor secret. The ugly (and alternative-sequence-breaking) railings are gone, and instead of SR50+luck you have to find the way up to a much higher platform.

Allowed jumping from the semifinal puzzle platform into the guard tower areas because there was nothing showing you could not. It doesn't actually achieve anything besides letting you grab the ammo from the zombies inside.

Flipped a bunch of linedefs to address #996.

Also add a switch for that platform-gate since its non-orthogonal shape and large size can't help but signal that it is expected to be remote-lowered.