levels: misc. E3M1 improvements.
Un-doomified the start:
- Hell Keep: start outdoors, touch freaky flesh thing, fight imps, find red brick gate, grab ammo boxes by door, hit red brick skull switch, receive immediate cacodemon.
- Land of the Lost: start outdoors, no freaky flesh thing, fight serpentipedes, find red brick gate, grab ammo boxes by door, hit red brick skull switch, receive immediate trilobite.
- Newer, Better Land of the Lost: start outdoors, fight serpentipedes, find red brick building, *shoot* freaky flesh thing, reveal corridor with ammo boxes and *slightly* less immediate trilobite.

Replaced candelabras with evil eyes for the shooting messaging.

Made exit teleporter lines monster blocking.

It is impossible to get through that mandatory northwest area without taking damage from the hurtfloor. Added a catwalk. All the health bonuses are on that catwalk now.

Made the lowering rock wall use the same grey floor as the surrounding area. Now all red surfaces you can reach without jumping are hurtfloors. Its trigger is now the instant-move lowest-ceiling trick so you can't race over to watch it move.

Two stimpacks are now medikits, while the number of medikits in the exit room are reduced to one. This increases the total health by 5 points (in addition to the 5 for losing the obligatory damagefloor) while reducing the likelihood of the player running into the exit room and (unintentionally) turning what should be a dynamic fight around the map into a chokepoint slog.

Broke all the fake contrast orthogonal lines in the ribbed pillars.