levels: minor e2m4 fixes.
Berserk pack secret sector was too narrow and could sometimes lets you grab the items without triggering the secret. That closet is now deepened (and slightly redecorated to make up for the new blank space).

The area around the waterfall was too dark compared to everything around it. Its light is now all 144 except right next to the light of the silver arch.

The horizontal trim on the silver arches now go all the way around.