levels: minor E3 improvements.
E3M1
After some discusssion on Discord it was agreed that the trilobite should not be visible until the player has discovered the berserk. The first trilobite is now outside of that hall, its back to the player looking at the berserk, and the hall has some sound blocker lines so it doesn't react if you try to plink at the worms (despite the berserk being right there).

I've also done a bit of redecorating to the front steps since we no longer need the evil eyes: the entire stairway is lined with candles and the eyes/ candelabras are replaced with blood fountains (as too many lights here would imply an area much brighter than what we're dealing with).

E3M3
The starting area lavafall had an unintentional fake contrast. This lavafall is now made more complex, with intentional fake contrasts.