levels: more minor fixes. (#1063)
* levels: one more texture fix on that Map17 ramp.

* levels: one more minor Map17 METAL2 alignment.

On the ledge to the right of the yellow door - the seam of the texture actually should not follow the ledge of the path, but continue the tiling pattern.

* levels: fix map03 start room texture.

See PR for screenshots, this one was really hard to see before the new METAL2 textures.

* levels: relocate e3m3 secret missile launcher.

Where it was positioned it was possible to grab it - even sometimes see it in vanilla - before that door even opens.

* levels: move e3m3 start arena hanging corpses.

They're supposed to be hanging from the bottom of the wooden rim, but sometimes vanilla errs in favour of sticking the actor to the higher ceiling in ambiguous rounding-error situations. The centre of each of these actors is now totally unambiguously inside the wooden rim sector.

* levels: address #1053.

* levels: address #1053 properly.

Also un-merge a bunch of sectors because it turns out blocking sounds from certain areas is actually kinda important.

* levels: fix Map31 BFG.

Moved it slightly closer to the teleorter pad so it's less possible to jump over it entirely by mistake, and aligned the textures in there.

* levels: more map31 adjustments.

- Re-merged the doorway sectors so any single gunshot will alarm all tripods.
- Some more texture fixes, replacing a few ceilings that didn't tile properly.
- More adjustments to the squeeze-past chamber, mostly just for texture alignment purposes.
- Made *all* the gore in the switch room point to the switch with the red card.
- Shortened the supersecret door switch so the gargoyle wouldn't repeat like that.

* levels: bring map31 secret closer, but conceal it.

* levels: fix e2m1 easy hot start.

* levels: misc. E3M1 improvements.

Un-doomified the start:
- Hell Keep: start outdoors, touch freaky flesh thing, fight imps, find red brick gate, grab ammo boxes by door, hit red brick skull switch, receive immediate cacodemon.
- Land of the Lost: start outdoors, no freaky flesh thing, fight serpentipedes, find red brick gate, grab ammo boxes by door, hit red brick skull switch, receive immediate trilobite.
- Newer, Better Land of the Lost: start outdoors, fight serpentipedes, find red brick building, *shoot* freaky flesh thing, reveal corridor with ammo boxes and *slightly* less immediate trilobite.

Replaced candelabras with evil eyes for the shooting messaging.

Made exit teleporter lines monster blocking.

It is impossible to get through that mandatory northwest area without taking damage from the hurtfloor. Added a catwalk. All the health bonuses are on that catwalk now.

Made the lowering rock wall use the same grey floor as the surrounding area. Now all red surfaces you can reach without jumping are hurtfloors. Its trigger is now the instant-move lowest-ceiling trick so you can't race over to watch it move.

Two stimpacks are now medikits, while the number of medikits in the exit room are reduced to one. This increases the total health by 5 points (in addition to the 5 for losing the obligatory damagefloor) while reducing the likelihood of the player running into the exit room and (unintentionally) turning what should be a dynamic fight around the map into a chokepoint slog.

Broke all the fake contrast orthogonal lines in the ribbed pillars.

* levels: minor e2m4 fixes.

Berserk pack secret sector was too narrow and could sometimes lets you grab the items without triggering the secret. That closet is now deepened (and slightly redecorated to make up for the new blank space).

The area around the waterfall was too dark compared to everything around it. Its light is now all 144 except right next to the light of the silver arch.

The horizontal trim on the silver arches now go all the way around.

* levels: minor E3 improvements.

E3M1
After some discusssion on Discord it was agreed that the trilobite should not be visible until the player has discovered the berserk. The first trilobite is now outside of that hall, its back to the player looking at the berserk, and the hall has some sound blocker lines so it doesn't react if you try to plink at the worms (despite the berserk being right there).

I've also done a bit of redecorating to the front steps since we no longer need the evil eyes: the entire stairway is lined with candles and the eyes/ candelabras are replaced with blood fountains (as too many lights here would imply an area much brighter than what we're dealing with).

E3M3
The starting area lavafall had an unintentional fake contrast. This lavafall is now made more complex, with intentional fake contrasts.

* levels: final E2M4 touchups by Xindage.

* levels: make map06 exit pad line monster blocking.

Also unmerge the sectors on two sides of a door that was causing some monsters to wake prematurely.

* levels: fix textures by map11 warehouse counter.

that new METAL2 sure makes a lot of things more visible...