sprites: Improve the skull on the skull column (#1090)
The skull that was here before was kind of a joke from the early days of the project, when it was the only skull graphic we had and used it everywhere to ridiculous extremes. Here I've replaced it with a shrunk-down version of the M_SKULL graphic used on the main menu, with some minor tweaks (the skull looks sadder than the menu one).

The new sprite is two pixels taller than the old one.