manual/dehacked: edit for clarity.
Some probably overambitious attempts to address #992.

[Note: I do not speak Spanish. Some longer translations are copied almost verbatim from DeepL output. Efforts have been made to parse the basic grammar of the output before committing anything but idiom is by no means guaranteed. If you find *anything* odd (that isn't obviously motivated by alliteration for the IDBEHOLD cheat), or any item/monster names already established that I'm not using here, please go ahead and fix it or tell someone who will! (Or tell me in the replies to this PR if you manage to catch it that soon)]

**General**

Rephrased the "Installing Freedoom" to make it a little easier on the kind of person who would actually need it: vanilla compatibility is also implied. FreeDM is also mentioned.

The armour example has been rewritten so we're only dealing with amounts the player will lose, instead of making the reader do internal arithmetic to track what went where.

The colorblind accessibility section for the keys is now just part of the description of the key pickups, and is also present in the Spanish manual. The keys are presented in the same order as they appear on the status bar and I've added a mnemonic that I found helpful.

Loading quicksaves is clarified: no need to die and menu is still available.

Miscellaneous rewordings to make things more to the point and avoid overexplaining.

Removed some references to "higher-level" and "lower-level" monsters. Someone who actually needs to be told this stuff might think this was an RPG thing.

Converted some CxMy references in the cheat code section.

**Ammo**

It's needlessly confusing to have "ammo" be used for both the pistol/minigun ammo and ammo generally. The gun pedant in me died a little inside, but the pickup messages now say "bullets" in line with both "bullets" and "balas" in the manuals.

(I've floated the idea of replacing the ammo box sprite with something without a label, but that's fairly low priority since everyone can see themselves touching the green square thing and only seeing one number go up.)

Clarified that backpacks don't stack their carrying capacity increase.

**Weapons**

Spanish weapon descriptions now bold the name as they are bolded in English.

Handgun description now uses the pickup sprite.

Pump shotgun's spread is clearly neither tight nor long for anyone who is playing at higher resolutions than vanilla 320x200. And it's not the main workhorse weapon in Phase 2.

Double-barreled shotgun isn't "twice" as powerful as the pump by any measurement. (Even its rate of shell consumption over time is more than double.) The much dimmer and unorthodoxically distributed red flash in #1085 can provide an excuse for the damage disparity for now.

SKAG no longer implies the ball does radius damage.

**Monsters**

The animal names are an evocative metaphor but the poetry is lost in doubling down on them in the descriptions when every single one is clearly taxonomically not that thing. ("Worm" was left untouched however since really that could mean any long wiggly thing.)

Matribite and hatchling image alt text should be capitalized for consistency; "hatchling" de-captalized in matribite description.

Painlord Spanish description seemed to imply it could only survive 5 missile *explosions* not direct hits, whereas the English is a bit clearer by using the same word for both painbros.

Large technospider and tripod explicitly said to be immune to explosions (but not SKAG).