music: use tremolo strings for d_map19 not piano. (#1080)
When I first started up Map19 with music for the first time in ages I thought the MIDI had glitched somehow and the instrumentation fell back on "Acoustic Grand" the way a lot of things did when I was younger. When I checked the MIDI it still sounded like it was supposed to be something else and the composer forgot to set the instrument, or it's a *really* experimental thing that frankly doesn't really fit a game - or at least not something that is not itself an extremely experimental abstract concept level which Map19 is not.

After some experimentation tremolo strings stood out by a fairly large margin as the best option.