dehacked: revise E4 endtext.
This started as an attempt to think through how to make the last stretch of E4M8 look a little more like the inside of a spaceship, when it occurred to me that it's very possible to end up with a lot of monsters still alive inside what appears on the map to be the body of the spaceship - so what happens with them? How do you relax with them banging down the door the whole way home?

This describes them fleeing the ship in the immediate aftermath of hitting the exit switch.