dehacked: adjust obituaries.
First, let's just make all the DM ones consistent in naming the victim first, so it's easier to quickly skim in the middle of a match.

I've also fixed the comment now that these are compatible with not just ZDoom.

There's a couple things that are inherited from ZDoom that don't even make sense, or are clearly derived from the Doom monsters - I don't think anything in FD calls the painlords "bruisers".

The "smitten" never sat well with me because the primary usage I see is like "she's so smitten with him, just look at all the little hearts she drew on his yearbook photo" and not like smite as in lightning.

Monsters with (ostensibly) player-equivalent weapons are given equivalent obituaries. (I love how deadpan the zombieman "killed" is and I think it works for PvP too.)

Slashed didn't even make sense for the id imp.

Both worms deserve to eat.

I'm sorry but we *have* to make that trilo-bite pun.

I've taken the liberty to move "perforate" to the shotgun to reflect how it tends to make a very visibly straight line of hits compared to the other bullet weapons. (Technically pistol and minigun do this too but it's harder to see as they're not all at the same time and the puff moves upwards.)

Yeah, I changed the railgun. Was anyone ever actually comfortable with that innuendo?

P.E.W. was a forced meme anyway. :(