sprites: Shotgun redraws/recolours by Korp. (#1085)
PASG HUD sprites are Blastfrog's originals with significant modifications.

DBSG HUD sprites are Fredrik's originals with not *as* significant modifications but I believe the wires are new. (Did Rei do those?)

PASG pickup sprite is my edit of the current one. (Removed the stock because the guns give a lesser sense of recoil now, and the pump animation makes way more sense if you assume the gun is not shouldered; also I didn't feel like adding a stock to the DBSG pickup for consistency.)

DBSG pickup sprite is my edit of Z86's original.