levels: tweak Map23 exit.
This is a mandatory damage exit that could kill someone even if they do everything else (more or less) right in beating this map. Worst case scenario, they "learn" from their "mistake", "realize" this is a "fake exit", and spend forever trying to find the "real" exit before giving up and learning much later that they'd actually found it before.

- You can pick up at least 30 health right next to the exit teleporter pad.
- The actual teleport landing itself is not a hurtfloor. It is, however, extremely narrow and right in front of an ambushing trilobite. The only way to dodge its fireball is down.
- Made the exit portal huge and fireblu and not a St. Peter's cross (which is often used elsewhere as a purely decorative thing).
- The "EXIT" map lines are now actually faintly glowing on the floor.

As an additional bonus edit: if you exit without finding that creepy area in the facility and have all those guys warp in, those un-warped guys are now all free to roam the entire area as soon as you fall off that high ledge. This means the necromancers can go into that crowd of zombies and raise them, as well as close range to attack you. Enjoy this epic battle - or panicked rush for the exit.