levels: minor fixes to E2M7.
Armour bonuses still floating on sector 599. They now adorn the side pillars instead.

The minigun is comically easy to miss - pretty much every passage into that area is designed for you to look the other way or overshoot it jumping in there, and you might go through the next 80% of the map before having any reason to go back into that part of the room. Its platform has been moved to the lower area, with the upper platform being ostensibly a shotgun but the shotgun guy now standing there took it.